Positioned featured using margin and padding
For various viewport widths, the margin and padding of the
featured-image and featured-section are adjusted accordingly to line up.

This method may present a problem if the height of the news-section
changes due to the visitor selecting various news posts.

For tablet-sized devices (≥768px) there remains a large empty whitespace
between the intro and the featured image.