Moved the featured image and section upward
The columns are offset using margin-top of -50px. Since the offset is
arbitrary and constant, this solution may not look correct when the
height of the news section changes. For example, if the news is the same
height as the intro, then the featured image may overlap the intro text.