Added circular background to the robot image
Since the background is a part of the image itself, the background and
the robot scale together easily without having to define the circle's
width and height and using absolute positioning. This also removes the
need for border-radius, which is not supported in IE 7.

One disadvantage is that the color, shape, and size of the background is
not easily modifiable, as the image itself must be edited. Additional
featured items would also require the same embedded background.