Added selection color
Since IE 8 does not support rgba values, the hex value is kept.

In Firefox, text selection only highlights the text and not the space in
between lines as in Chrome. Horizontal rules are also not highlighted.
Additionally, selecting an image does not follow the color specified in
the ::-moz-selection rule.