Added more spacing between last element and footer
Depending on the viewport width, the last element next to the footer is
either the featured text or the featured image. By default, bootstrap
applies a 10px margin to the bottom of all p elements, thus causing the
spacing between the last element and the footer to be inconsistent.

To deal with this the bottom margin is removed from the featured text.
This allows the spacing before the footer to be defined using the body >
.container selector and changing the padding.

A better solution may be to remove the bottom margin from all
terminating p elements using last of selector and place in common
styles, or remove the margin from all p elements and manually space p
elements and divs/sections.