Added margin to the top adjacent p elements
If a p element is the last element before the footer, then the space in
between the two changes when switching from the smallest viewport layout
to the larger ones. It would be better is the space is consistent and
defined in one style rule.

Instead of having the margin at the bottom of all p elements (bootstrap
default), the margin is placed at the top of adjacent p elements. This
rule may have to be added to other elements such as lists if they are
the last elements before the footer.