Added group of robot images to work page
The three robot images are wrapped in a div which uses max-width of 40%
and floated to the right using pull-right. The secondary images are
placed alongside each other using display inline-block and the right one
is set to pull-right. Their size is restricted using max-width 46%. All
three images use the img-responsive class.

The bootstrap thumbnail class is not used since it does not work
correctly with the max-width properties.