Added icon fonts for footer links
Added a subset of the Font Awesome icon font for the footer links using
fontello.com. Since the subset does not include the Font Awesome styles,
the sizing classes such as fa-lg, fa-2x, etc. are not available.

There is some extra space that is introduced below the footer so that it
is not sticking directly to the bottom.