Floated work images to the right of main text
Instead of the images floating to the right of the tables, they
float to the right of the article text

While this method works okay for phone screen sizes (although it makes
the images very small) it leaves a large gap between the hr and the
image for larger viewports

A better approach would be to use this layout for phone views, then move
the images back by the tables for larger viewports. The ordering of the elements may be
changed using jQuery.