Floated work images left and right
Did not use Bootstrap class img-responsive since it sets max-width to
100%. Instead just set max-width to 40% and height to auto.

For the robot article, since the last paragraph is short, the floated
robot poster leaves a large space for wider viewports.

Should reduce the size of the robot image and QA poster since they are
too large. For the widest viewport they cut down into the paragraph
below.