Added horizontal layout for robot images
For medium and large devices, the floated images are hidden and a
separate horizontal row of the same images is shown. The images are
placed in a nested grid and the Bootstrap class hidden-* is used.

There is an issue with the anchor tag being extended outside of the
image borders. This is most likely due to the center-block class. Can
try using thumbnail class or using an unordered list instead of a grid
to group the images.