Added profile picture using grid system
Added two different sized profile pictures to about page. The smaller is
shown in a stacked column layout for narrow viewports. The other is
shown in two column layout for larger viewports.

For narrow viewports, there is a lot of extra space to left and right of
the picture because of the extra small viewport breakpoint. May have to
change the breakpoint width, float the image, or change its size.

NOTE: Since Bootstrap 3.2.0, visible-xs is deprecated. Will have to use
visible-xs-block instead.