Added vertical work images with img-responsive
With this solution the images reach their maximum size for larger
viewports, thus leaving large margins on the sides. On narrow viewports
the images extend to the width of the container.

Scaled QA poster using Lanczos interpolation since it slightly reduces
the errors around the text.

Increased the minimum viewport width required for viewing images with
Magnific since the images are wider in this layout and even more space
is required for viewing the text of the posters.