Removed clickable margins of responsive images
Using the Bootstrap class "img-resonsive" for images inside anchors
causes the margins to be clickable/linkable. This is because it applies
display:block to the image (for horizontally centering), which
essentially makes the anchors block level as well, thereby extending the
margins to the width of the parent.

By only applying max-width:100% and height:auto to the images and
wrapping the anchors in a div and applying text-align:center, the images
are made responsive and horizontally centered. This allows for the
margins to not be clickable.