Fixed width of thumbnail for QA poster in IE8
In IE8, the width of the QA poster thumbnail would stretch to the entire
width of the page because the image is responsive and the width and
height attributes did not match the actual size of the image. This was
fixed by including a properly downsized thumbnail (512x384px).

This solution is semantically more correct and results in faster page
loads for the user since only the low-res thumbnail is downloaded
first. The full resolution image is downloaded only if the user wishes
to see it. Prior to this, the full-res was used for both thumbnail and
the full size link.

The width and height attributes were Initially manipulated in order to
have the browser scale the image since this resulted in higher quality
renderings compared to downsizing the image. The problem was that the
resizing was being performed from the 960x720px JPEG, which contains
ringing and aliasing around the text.

The thumbnail in this commit is exported directly from the Powerpoint
as a PNG.